refactor(mechanics): update name files
